Move aggregate query logic to GridTable to pass to GridTableWidgets and GridTableCharts
