+++ atd-vze/src/Components/GridTableCharts.js
@@ -1,0 +1,111 @@
+import React, { useState, useEffect } from "react";
+
+import { Col } from "reactstrap";
+
+const GridTableCharts = ({}) => {
+  const { count: crashCount } = data.atd_txdot_crashes_aggregate.aggregate;
+
+  const vehBodyGraphConfig = {
+    palette: palette(
+      "mpn65",
+      data.atd_txdot_locations[0].crashes_by_veh_body_style.length
+    ).map(a => `#${a}`),
+    labels: data.atd_txdot_locations[0].crashes_by_veh_body_style.map(
+      a => a.veh_body_styl_desc
+    ),
+    data: data.atd_txdot_locations[0].crashes_by_veh_body_style.map(
+      a => a.count
+    ),
+  };
+
+  const doughnut = {
+    labels: vehBodyGraphConfig.labels,
+    datasets: [
+      {
+        data: vehBodyGraphConfig.data,
+        backgroundColor: vehBodyGraphConfig.palette,
+        hoverBackgroundColor: ["#000"],
+      },
+    ],
+  };
+
+  const horizontalBar = {
+    labels: data.atd_txdot_locations[0].crashes_by_manner_collision
+      .map((a, index) => `${index + 1}. ${a.collsn_desc}`)
+      .map(a => formatLabel(a)),
+    datasets: [
+      {
+        label: "Number of Collisions",
+        backgroundColor: colors.success,
+        borderColor: colors.grey200,
+        borderWidth: 1,
+        hoverBackgroundColor: colors.darkgreen,
+        hoverBorderColor: colors.grey700,
+        data: data.atd_txdot_locations[0].crashes_by_manner_collision.map(
+          a => a.count
+        ),
+      },
+    ],
+  };
+
+  return (
+    <Col>
+      <Card>
+        <CardHeader>Types of Vehicles - Count Distribution</CardHeader>
+        <CardBody>
+          {crashCount === 0 && (
+            <Alert color="warning">
+              No crashes at this particular location
+            </Alert>
+          )}
+
+          {crashCount > 0 && (
+            <div className="chart-wrapper" style={{ padding: "1.5rem 0" }}>
+              <Badge
+                color="dark"
+                className="float-right"
+                style={{ padding: "4px" }}
+              >
+                <i className="fa fa-mouse-pointer" />
+                &nbsp; Click On Labels
+              </Badge>
+              <Doughnut data={doughnut} />
+            </div>
+          )}
+        </CardBody>
+      </Card>
+      <Card>
+        <CardHeader>Manner of Collisions - Most Frequent</CardHeader>
+        <CardBody>
+          {crashCount === 0 && (
+            <Alert color="warning">
+              No crashes at this particular location
+            </Alert>
+          )}
+
+          {crashCount > 0 && (
+            <div className="chart-wrapper" style={{ padding: "1.5rem 0" }}>
+              <HorizontalBar
+                data={horizontalBar}
+                options={{
+                  scales: {
+                    xAxes: [
+                      {
+                        ticks: {
+                          beginAtZero: true,
+                          precision: 0,
+                        },
+                      },
+                    ],
+                  },
+                }}
+              />
+            </div>
+          )}
+        </CardBody>
+      </Card>
+    </Col>
+  );
+};
+
+export default GridTableCharts;
