+++ atd-vze/src/Components/GridTableWidgets.js
@@ -1,30 +1,10 @@
 import React, { useState, useEffect } from "react";
 import { useLazyQuery } from "@apollo/react-hooks";
 
-import { Container, Col, Row } from "reactstrap";
+import { Col, Row } from "reactstrap";
 import Widget02 from "../views/Widgets/Widget02";
 
-const GridTableWidgets = ({ query, queryConfig, widgetsConfig }) => {
-  const [aggregateQuery, setAggregateQuery] = useState(null);
-
-  const [loadAggData, { loading: aggLoading, data: aggData }] = useLazyQuery(
-    aggregateQuery
-  );
-
-  // Update aggregate query with latest filters
-  useEffect(() => {
-    const aggregatesQuery = query.queryAggregate(queryConfig, query);
-    // Prevent endlessly setting aggregateQuery
-    if (aggregatesQuery !== aggregateQuery) {
-      setAggregateQuery(aggregatesQuery);
-    }
-  });
-
-  // Execute aggregate query each time it changes
-  useEffect(() => {
-    aggregateQuery !== null && loadAggData();
-  }, [aggregateQuery]);
-
+const GridTableWidgets = ({ aggData, widgetsConfig }) => {
   const getAggregateData = field => {
     // If aggData is populated, use switch to return aggregate data
     if (aggData !== undefined && Object.keys(aggData).length > 0) {
@@ -68,7 +48,8 @@ const GridTableWidgets = ({ query, queryConfig, widgetsConfig }) => {
           <Col key={i} xs="12" sm="6" md="4">
             <Widget02
               key={i}
-              header={getAggregateData(widget.getAggregateDataArg)}
+              // DB sometimes return null so display "0" in that case
+              header={getAggregateData(widget.getAggregateDataArg) || "0"}
               mainText={widget.mainText}
               icon={widget.icon}
               color={widget.color}
