+++ atd-vze/src/views/Locations/Location.js
@@ -76,36 +76,6 @@ function Location(props) {
     ),
   };
 
-  const doughnut = {
-    labels: vehBodyGraphConfig.labels,
-    datasets: [
-      {
-        data: vehBodyGraphConfig.data,
-        backgroundColor: vehBodyGraphConfig.palette,
-        hoverBackgroundColor: ["#000"],
-      },
-    ],
-  };
-
-  const horizontalBar = {
-    labels: data.atd_txdot_locations[0].crashes_by_manner_collision
-      .map((a, index) => `${index + 1}. ${a.collsn_desc}`)
-      .map(a => formatLabel(a)),
-    datasets: [
-      {
-        label: "Number of Collisions",
-        backgroundColor: colors.success,
-        borderColor: colors.grey200,
-        borderWidth: 1,
-        hoverBackgroundColor: colors.darkgreen,
-        hoverBorderColor: colors.grey700,
-        data: data.atd_txdot_locations[0].crashes_by_manner_collision.map(
-          a => a.count
-        ),
-      },
-    ],
-  };
-
   const { count: crashCount } = data.atd_txdot_crashes_aggregate.aggregate;
 
   return (
@@ -201,62 +171,7 @@ function Location(props) {
             );
           })}
         </Col>
-        <Col md="6">
-          <Card>
-            <CardHeader>Types of Vehicles - Count Distribution</CardHeader>
-            <CardBody>
-              {crashCount === 0 && (
-                <Alert color="warning">
-                  No crashes at this particular location
-                </Alert>
-              )}
-
-              {crashCount > 0 && (
-                <div className="chart-wrapper" style={{ padding: "1.5rem 0" }}>
-                  <Badge
-                    color="dark"
-                    className="float-right"
-                    style={{ padding: "4px" }}
-                  >
-                    <i className="fa fa-mouse-pointer" />
-                    &nbsp; Click On Labels
-                  </Badge>
-                  <Doughnut data={doughnut} />
-                </div>
-              )}
-            </CardBody>
-          </Card>
-          <Card>
-            <CardHeader>Manner of Collisions - Most Frequent</CardHeader>
-            <CardBody>
-              {crashCount === 0 && (
-                <Alert color="warning">
-                  No crashes at this particular location
-                </Alert>
-              )}
-
-              {crashCount > 0 && (
-                <div className="chart-wrapper" style={{ padding: "1.5rem 0" }}>
-                  <HorizontalBar
-                    data={horizontalBar}
-                    options={{
-                      scales: {
-                        xAxes: [
-                          {
-                            ticks: {
-                              beginAtZero: true,
-                              precision: 0,
-                            },
-                          },
-                        ],
-                      },
-                    }}
-                  />
-                </div>
-              )}
-            </CardBody>
-          </Card>
-        </Col>
+        <Col md="6"></Col>
       </Row>
       <Row>
         <Col>
